lbs::XXPowerIterationKEigen is in 1 .h/.cc file
